refactor: split Api
